cdb2api: One definition of ssl_init
This fixes the following linker error
ld: fatal: symbol '$XAaxlqFBwqC9JDwILy4DWrGJXCb2$BMC.is_ssl_initialized' is multiply-defined:
    (file util/libutil.a(ssl_support.c.o) type=OBJT; file cdb2api/libopencdb2api.a(ssl_support.c.o) type=OBJT);
