+++ util/ssl_support.c
@@ -52,126 +52,6 @@
 #define my_ssl_eprintln(fmt, ...)    \
     ssl_eprintln("Generic", "%s: " fmt, __func__, ##__VA_ARGS__)
 
-static int is_ssl_initialized = 0;
-static pthread_mutex_t ssl_init_lock = PTHREAD_MUTEX_INITIALIZER;
-
-#ifndef CRYPTO_num_locks
-/* Callbacks for OpenSSL locking. OpenSSL >= 1.1.0 has its own locking.
-   CRYPTO_num_locks is a function in OpenSSL < 1.1.0 but is made a macro
-   in OpenSSL >= 1.1.0. So we assume that if CRYPTO_num_locks is not
-   defined, we still need to implement our own locking. */
-static pthread_mutex_t *ssl_locks = NULL;
-
-#ifdef OPENSSL_NO_DEPRECATED
-static void ssl_threadid(CRYPTO_THREADID *thd)
-{
-    CRYPTO_THREADID_set_numeric(thd, (intptr_t)pthread_self());
-}
-#endif /* OPENSSL_NO_DEPRECATED */
-
-/* For OpenSSL < 1.0.0. */
-static unsigned long ssl_threadid_deprecated()
-{
-    return (unsigned long)pthread_self();
-}
-
-static void ssl_lock(int mode, int type, const char *file, int line)
-{
-    int rc;
-    if (mode & CRYPTO_LOCK) {
-        if ((rc = pthread_mutex_lock(&ssl_locks[type])) != 0)
-            my_ssl_eprintln("Failed to lock pthread mutex: rc = %d.", rc);
-    } else {
-        if ((rc = pthread_mutex_unlock(&ssl_locks[type])) != 0)
-            my_ssl_eprintln("Failed to unlock pthread mutex: rc = %d.", rc);
-    }
-}
-#endif /* CRYPTO_num_locks */
-
-int SBUF2_FUNC(ssl_init)(int init_openssl, int init_crypto,
-                       int lock, char *err, size_t n)
-{
-    int rc;
-
-    if (init_openssl) {
-#ifndef OPENSSL_THREADS
-        /* OpenSSL is not configured for threaded applications. */
-        ssl_sfeprint(err, n, my_ssl_eprintln,
-                     "OpenSSL is not configured with thread support.");
-        return EPERM;
-#endif /* OPENSSL_THREADS */
-    }
-
-	rc = lock ? pthread_mutex_lock(&ssl_init_lock) : 0;
-    if (rc != 0)
-        return rc;
-
-    /* Initialize OpenSSL only once. */
-    if (!is_ssl_initialized) {
-#ifndef CRYPTO_num_locks
-        if (init_crypto) {
-            /* Configure SSL locking.
-               This is only required for OpenSSL < 1.1.0. */
-            int nlocks = CRYPTO_num_locks();
-            ssl_locks = malloc(sizeof(pthread_mutex_t) * nlocks);
-            if (ssl_locks == NULL) {
-                ssl_sfeprint(err, n, my_ssl_eprintln,
-                             "Failed to allocate SSL locks.");
-                rc = ENOMEM;
-                goto error;
-            }
-
-            int ii;
-            for (ii = 0; ii < nlocks; ++ii) {
-                if ((rc = pthread_mutex_init(&ssl_locks[ii], NULL)) != 0) {
-                    /* Whoops - roll back all we have done. */
-                    while (ii > 0) {
-                        --ii;
-                        pthread_mutex_destroy(&ssl_locks[ii]);
-                    }
-                    free(ssl_locks);
-                    ssl_locks = NULL;
-                    ssl_sfeprint(err, n, my_ssl_eprintln,
-                                 "Failed to initialize mutex: %s", strerror(rc));
-                    goto error;
-                }
-            }
-#  ifdef OPENSSL_NO_DEPRECATED
-            CRYPTO_THREADID_set_callback(ssl_threadid);
-#  else
-            /* Use deprecated functions for OpenSSL < 1.0.0. */
-            CRYPTO_set_id_callback(ssl_threadid_deprecated);
-#  endif /* OPENSSL_NO_DEPRECATED */
-            CRYPTO_set_locking_callback(ssl_lock);
-        }
-#endif /* CRYPTO_num_locks */
-
-       /* Configure the library. */
-        if (init_openssl) {
-#if OPENSSL_VERSION_NUMBER >= 0x10100000L
-            OPENSSL_init_ssl(OPENSSL_INIT_LOAD_CONFIG, NULL);
-#else
-            OPENSSL_config(NULL);
-            SSL_library_init();
-            SSL_load_error_strings();
-#endif /* OPENSSL_VERSION_NUMBER */
-        }
-        is_ssl_initialized = 1;
-    }
-
-#ifndef CRYPTO_num_locks
-error:
-#endif /* CRYPTO_num_locks */
-    if (lock) {
-        if (rc == 0)
-            rc = pthread_mutex_unlock(&ssl_init_lock);
-        else
-            pthread_mutex_unlock(&ssl_init_lock);
-    }
-
-    return rc;
-}
-
 #if SBUF2_SERVER
 /* server-side session id context */
 static unsigned char sid_ctx[8];
